BPE26-ref: extract volume feature to dedicated file
