+++ src/features/ui/volume.ts
@@ -1,0 +1,86 @@
+import { PLUME_CONSTANTS, PLUME_ELEM_IDENTIFIERS } from "../../domain/plume";
+import { getPlumeUiInstance, PLUME_ACTION_TYPES } from "../../infra/AppInstanceImpl";
+import { getStoreInstance, STORE_ACTION_TYPES } from "../../infra/AppStoreImpl";
+import { PLUME_SVG } from "../../svg/icons";
+import { getString } from "../i18n";
+
+const { VOLUME_SLIDER_GRANULARITY } = PLUME_CONSTANTS;
+
+export function syncMuteBtn(isMuted: boolean): void {
+  const plumeUiInstance = getPlumeUiInstance();
+  const plume = plumeUiInstance.getState();
+  const newMuteBtn = plume.muteBtn;
+
+  newMuteBtn.innerHTML = isMuted ? PLUME_SVG.volumeMuted : PLUME_SVG.volumeOn;
+  newMuteBtn.title = isMuted ? getString("ARIA__UNMUTE") : getString("ARIA__MUTE");
+  newMuteBtn.ariaLabel = isMuted ? getString("ARIA__UNMUTE") : getString("ARIA__MUTE");
+  newMuteBtn.ariaPressed = isMuted.toString();
+  newMuteBtn.classList.toggle("muted", isMuted);
+  plumeUiInstance.dispatch({ type: PLUME_ACTION_TYPES.SET_MUTE_BTN, payload: newMuteBtn });
+}
+
+export function handleMuteToggle(): void {
+  const store = getStoreInstance();
+  store.dispatch({ type: STORE_ACTION_TYPES.TOGGLE_MUTE });
+}
+
+export async function createVolumeControlSection(): Promise<HTMLDivElement | null> {
+  const store = getStoreInstance();
+  const plumeUiInstance = getPlumeUiInstance();
+  const plume = plumeUiInstance.getState();
+
+  if (plume.volumeSlider) return null;
+
+  const container = document.createElement("div");
+  container.id = PLUME_ELEM_IDENTIFIERS.volumeContainer.split("#")[1];
+
+  const muteBtn = document.createElement("button");
+  muteBtn.id = PLUME_ELEM_IDENTIFIERS.muteBtn.split("#")[1];
+  muteBtn.type = "button";
+  muteBtn.title = getString("ARIA__MUTE");
+  muteBtn.ariaLabel = getString("ARIA__MUTE");
+  muteBtn.ariaPressed = "false";
+  muteBtn.innerHTML = PLUME_SVG.volumeOn;
+
+  const volumeSlider = document.createElement("input");
+  volumeSlider.id = PLUME_ELEM_IDENTIFIERS.volumeSlider.split("#")[1];
+  volumeSlider.type = "range";
+  volumeSlider.min = "0";
+  volumeSlider.max = VOLUME_SLIDER_GRANULARITY.toString();
+  const currentVolume = store.getState().volume;
+  volumeSlider.value = Math.round(currentVolume * VOLUME_SLIDER_GRANULARITY).toString();
+  volumeSlider.ariaLabel = getString("ARIA__VOLUME_SLIDER");
+
+  // Apply saved volume to audio element
+  plume.audioElement.volume = currentVolume;
+
+  const valueDisplay = document.createElement("div");
+  valueDisplay.id = PLUME_ELEM_IDENTIFIERS.volumeValue.split("#")[1];
+  valueDisplay.textContent = `${volumeSlider.value}${getString("META__PERCENTAGE")}`;
+
+  muteBtn.addEventListener("click", handleMuteToggle);
+
+  // Event listener for volume change via slider
+  volumeSlider.addEventListener("input", function (this: HTMLInputElement) {
+    const volume = Number.parseInt(this.value) / VOLUME_SLIDER_GRANULARITY;
+
+    plume.audioElement.volume = volume;
+    valueDisplay.textContent = `${this.value}${getString("META__PERCENTAGE")}`;
+
+    // Moving slider off zero counts as an intentional unmute
+    if (volume > 0 && store.getState().isMuted) {
+      store.dispatch({ type: STORE_ACTION_TYPES.SET_IS_MUTED, payload: false });
+    }
+
+    store.dispatch({ type: STORE_ACTION_TYPES.SET_VOLUME, payload: volume });
+  });
+
+  container.appendChild(muteBtn);
+  container.appendChild(volumeSlider);
+  container.appendChild(valueDisplay);
+
+  plumeUiInstance.dispatch({ type: PLUME_ACTION_TYPES.SET_VOLUME_SLIDER, payload: volumeSlider });
+  plumeUiInstance.dispatch({ type: PLUME_ACTION_TYPES.SET_MUTE_BTN, payload: muteBtn });
+
+  return container;
+}
